+++ qt/chip_db.h
@@ -32,17 +32,17 @@ public:
     ChipDb();
     virtual ~ChipDb();
 
-    int getParamFromHexString(const QString &value, uint32_t &param);
-    int getParamFromString(const QString &value, uint32_t &param);
+    int getParamFromHexString(const QString &value, quint64 &param);
+    int getParamFromString(const QString &value, quint64 &param);
     int getParamFromString(const QString &value, uint8_t &param);
-    int getStringFromParam(const uint32_t &param, QString &value);
-    int getHexStringFromParam(const uint32_t &param, QString &value);
-    int getOptParamFromString(const QString &value, uint32_t &param);
-    int getOptParamFromHexString(const QString &value, uint32_t &param);
-    int getStringFromOptParam(const uint32_t &param, QString &value);
-    int getHexStringFromOptParam(const uint32_t &param, QString &value);
-    bool isParamValid(uint32_t param, uint32_t min, uint32_t max);
-    bool isOptParamValid(uint32_t param, uint32_t min, uint32_t max);
+    int getStringFromParam(const quint64 &param, QString &value);
+    int getHexStringFromParam(const quint64 &param, QString &value);
+    int getOptParamFromString(const QString &value, quint64 &param);
+    int getOptParamFromHexString(const QString &value, quint64 &param);
+    int getStringFromOptParam(const quint64 &param, QString &value);
+    int getHexStringFromOptParam(const quint64 &param, QString &value);
+    bool isParamValid(quint64 param, quint64 min, quint64 max);
+    bool isOptParamValid(quint64 param, quint64 min, quint64 max);
 
     virtual QString getNameByChipId(uint32_t id1, uint32_t id2,
         uint32_t id3, uint32_t id4, uint32_t id5) = 0;
@@ -53,10 +53,10 @@ public:
     uint32_t pageSizeGetByName(const QString &name);
     uint32_t extendedPageSizeGetById(int id);
     uint32_t extendedPageSizeGetByName(const QString &name);
-    uint32_t totalSizeGetById(int id);
-    uint32_t totalSizeGetByName(const QString &name);
-    uint32_t extendedTotalSizeGetById(int id);
-    uint32_t extendedTotalSizeGetByName(const QString &name);
+    quint64 totalSizeGetById(int id);
+    quint64 totalSizeGetByName(const QString &name);
+    quint64 extendedTotalSizeGetById(int id);
+    quint64 extendedTotalSizeGetByName(const QString &name);
     void addChip(ChipInfo *chipInfo);
     void delChip(int index);
     int size();
@@ -69,8 +69,8 @@ public:
     int setPageSize(int chipIndex, uint32_t pageSize);
     uint32_t getBlockSize(int chipIndex);
     int setBlockSize(int chipIndex, uint32_t blockSize);
-    uint32_t getTotalSize(int chipIndex);
-    int setTotalSize(int chipIndex, uint32_t totalSize);
+    quint64 getTotalSize(int chipIndex);
+    int setTotalSize(int chipIndex, quint64 totalSize);
     uint32_t getSpareSize(int chipIndex);
     int setSpareSize(int chipIndex, uint32_t spareSize);
     uint8_t getBBMarkOffset(int chipIndex);
