+++ qt/parallel_chip_db.h
@@ -79,8 +79,8 @@ public:
         uint32_t id5);
     QString getNameByChipId(uint32_t id1, uint32_t id2,
         uint32_t id3, uint32_t id4, uint32_t id5) override;
-    uint32_t getChipParam(int chipIndex, int paramIndex);
-    int setChipParam(int chipIndex, int paramIndex, uint32_t paramValue);
+    quint64 getChipParam(int chipIndex, int paramIndex);
+    int setChipParam(int chipIndex, int paramIndex, quint64 paramValue);
 };
 
 #endif // PARALLEL_CHIP_DB_H
