+++ qt/parallel_chip_db_table_model.cpp
@@ -58,52 +58,52 @@ QVariant ParallelChipDbTableModel::data(const QModelIndex &index,
     case ParallelChipDb::CHIP_PARAM_BB_MARK_OFF:
         return chipDb->getBBMarkOffset(index.row());
     case ParallelChipDb::CHIP_PARAM_T_CS:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_CS);
     case ParallelChipDb::CHIP_PARAM_T_CLS:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_CLS);
     case ParallelChipDb::CHIP_PARAM_T_ALS:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_ALS);
     case ParallelChipDb::CHIP_PARAM_T_CLR:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_CLR);
     case ParallelChipDb::CHIP_PARAM_T_AR:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_AR);
     case ParallelChipDb::CHIP_PARAM_T_WP:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_WP);
     case ParallelChipDb::CHIP_PARAM_T_RP:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_RP);
     case ParallelChipDb::CHIP_PARAM_T_DS:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_DS);
     case ParallelChipDb::CHIP_PARAM_T_CH:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_CH);
     case ParallelChipDb::CHIP_PARAM_T_CLH:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_CLH);
     case ParallelChipDb::CHIP_PARAM_T_ALH:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_ALH);
     case ParallelChipDb::CHIP_PARAM_T_WC:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_WC);
     case ParallelChipDb::CHIP_PARAM_T_RC:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_RC);
     case ParallelChipDb::CHIP_PARAM_T_REA:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_T_REA);
     case ParallelChipDb::CHIP_PARAM_ROW_CYCLES:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_ROW_CYCLES);
     case ParallelChipDb::CHIP_PARAM_COL_CYCLES:
-        return chipDb->getChipParam(index.row(),
+        return (uint)chipDb->getChipParam(index.row(),
             ParallelChipInfo::CHIP_PARAM_COL_CYCLES);
     case ParallelChipDb::CHIP_PARAM_READ1_CMD:
         chipDb->getHexStringFromParam(chipDb->getChipParam(index.row(),
@@ -380,7 +380,7 @@ Qt::ItemFlags ParallelChipDbTableModel::flags (const QModelIndex &index) const
 bool ParallelChipDbTableModel::setData(const QModelIndex &index,
     const QVariant &value, int role)
 {
-    uint32_t paramVal;
+    quint64 paramVal;
 
     if (role != Qt::EditRole)
         return false;
