+++ qt/programmer.h
@@ -82,9 +82,9 @@ public:
     bool isHwEccEnabled();
     void setHwEccEnabled(bool isHwEccEnabled);
     void readChipId(ChipId *chipId);
-    void eraseChip(uint32_t addr, uint32_t len);
-    void readChip(SyncBuffer *buf, uint32_t addr, uint32_t len, bool isReadLess);
-    void writeChip(SyncBuffer *buf, uint32_t addr, uint32_t len,
+    void eraseChip(quint64 addr, quint64 len);
+    void readChip(SyncBuffer *buf, quint64 addr, quint64 len, bool isReadLess);
+    void writeChip(SyncBuffer *buf, quint64 addr, quint64 len,
         uint32_t pageSize);
     void readChipBadBlocks();
     void confChip(ChipInfo *chipInfo);
@@ -93,36 +93,36 @@ public:
     void firmwareUpdate(const QString &fileName);
 
 signals:
-    void connectCompleted(int ret);
-    void readChipIdCompleted(int ret);
+    void connectCompleted(quint64 ret);
+    void readChipIdCompleted(quint64 ret);
     void writeChipCompleted(int ret);
-    void writeChipProgress(unsigned int progress);
-    void readChipCompleted(int ret);
-    void readChipProgress(unsigned int ret);
-    void eraseChipCompleted(int ret);
-    void eraseChipProgress(unsigned int progress);
-    void readChipBadBlocksProgress(unsigned int progress);
-    void readChipBadBlocksCompleted(int ret);
-    void confChipCompleted(int ret);
+    void writeChipProgress(quint64 progress);
+    void readChipCompleted(quint64 ret);
+    void readChipProgress(quint64 ret);
+    void eraseChipCompleted(quint64 ret);
+    void eraseChipProgress(quint64 progress);
+    void readChipBadBlocksProgress(quint64 progress);
+    void readChipBadBlocksCompleted(quint64 ret);
+    void confChipCompleted(quint64 ret);
     void firmwareUpdateCompleted(int ret);
-    void firmwareUpdateProgress(unsigned int progress);
+    void firmwareUpdateProgress(quint64 progress);
 
 private slots:
-    void readChipIdCb(int ret);
+    void readChipIdCb(quint64 ret);
     void writeCb(int ret);
-    void writeProgressCb(unsigned int progress);
-    void readCb(int ret);
-    void readProgressCb(unsigned int progress);
-    void eraseChipCb(int ret);
-    void eraseProgressChipCb(unsigned int progress);
-    void readChipBadBlocksCb(int ret);
-    void readChipBadBlocksProgressCb(unsigned int progress);
-    void confChipCb(int ret);
+    void writeProgressCb(quint64 progress);
+    void readCb(quint64 ret);
+    void readProgressCb(quint64 progress);
+    void eraseChipCb(quint64 ret);
+    void eraseProgressChipCb(quint64 progress);
+    void readChipBadBlocksCb(quint64 ret);
+    void readChipBadBlocksProgressCb(quint64 progress);
+    void confChipCb(quint64 ret);
     void logCb(QtMsgType msgType, QString msg);
-    void connectCb(int ret);
-    void getActiveImageCb(int ret);
+    void connectCb(quint64 ret);
+    void getActiveImageCb(quint64 ret);
     void firmwareUpdateCb(int ret);
-    void firmwareUpdateProgressCb(unsigned int progress);
+    void firmwareUpdateProgressCb(quint64 progress);
 };
 
 #endif // PROGRAMMER_H
