+++ qt/reader.cpp
@@ -22,7 +22,7 @@ Reader::~Reader()
 }
 
 void Reader::init(const QString &portName, qint32 baudRate, SyncBuffer *rbuf,
-    uint32_t rlen, const uint8_t *wbuf, uint32_t wlen, bool isSkipBB,
+    quint64 rlen, const uint8_t *wbuf, uint32_t wlen, bool isSkipBB,
     bool isReadLess)
 {
     this->portName = portName;
