+++ qt/spi_chip_info.h
@@ -33,14 +33,14 @@ public:
 
 private:
     QByteArray halConf;
-    uint32_t params[CHIP_PARAM_NUM] = {};
+    quint64 params[CHIP_PARAM_NUM] = {};
 
 public:
     SpiChipInfo();
     virtual ~SpiChipInfo();
     const QByteArray &getHalConf() override;
-    uint32_t getParam(uint32_t num) override;
-    int setParam(uint32_t num, uint32_t value) override;
+    quint64 getParam(uint32_t num) override;
+    int setParam(uint32_t num, quint64 value) override;
 };
 
 #endif // SPI_CHIP_INFO_H
