+++ qt/writer.h
@@ -21,11 +21,11 @@ class Writer : public QObject
     QString portName;
     qint32 baudRate;
     SyncBuffer *buf;
-    uint32_t addr;
-    uint32_t len;
-    uint32_t pageSize;
-    uint32_t bytesAcked;
-    uint32_t bytesWritten;
+    quint64 addr;
+    quint64 len;
+    quint64 pageSize;
+    quint64 bytesAcked;
+    quint64 bytesWritten;
     bool skipBB;
     bool incSpare;
     bool enableHwEcc;
@@ -58,14 +58,14 @@ public:
     explicit Writer();
     ~Writer();
     void init(const QString &portName, qint32 baudRate, SyncBuffer *buf,
-        uint32_t addr, uint32_t len, uint32_t pageSize,
+        quint64 addr, quint64 len, uint32_t pageSize,
         bool skipBB, bool incSpare, bool enableHwEcc, uint8_t startCmd,
         uint8_t dataCmd, uint8_t endCmd);
     void start();
     void stop();
 signals:
     void result(int ret);
-    void progress(unsigned int progress);
+    void progress(quint64 progress);
     void log(QtMsgType msgType, QString msg);
 };
